urls pong changed
