+++ transcendence/base/templates/index.html
@@ -5,7 +5,7 @@
 			<div class="signup-header">
 				<h2 class="header-title">WELCOME TO THE PONG GAME!</h2>
 				<div class="buttons-container">
-					<button class="button1" onclick="loadContent('play/'); return false;" type="submit">PLAY</button>
+					<button class="button1" onclick="loadContent('pong/play/'); return false;" type="submit">PLAY</button>
 					<button class="button2" onclick="loadContent('user/profile/'); return false;" type="submit">PROFILE</button>
 				</div>
 			</div>
