+++ transcendence/base/templates/onevsone.html
@@ -48,12 +48,12 @@
 					type="text"
 					name="player2"
 					/>
-					<button class="button3" onclick="loadContent('play/'); return false;" type="submit">OK</button>
+					<button class="button3" onclick="loadContent('pong/play/'); return false;" type="submit">OK</button>
 				</div>
 			</div>
 		</div>
 		<div class="button3-container">
-			<button class="button4" onclick="loadContent('playonevsone/'); return false;" type="submit">LET'S FIGHT!!!</button>
+			<button class="button4" onclick="loadContent('pong/playonevsone/'); return false;" type="submit">LET'S FIGHT!!!</button>
 		</div>
 	</div>
 </div>
