+++ transcendence/base/templates/play.html
@@ -5,8 +5,8 @@
 			<div class="signup-header">
 				<h2 class="header-title">WHAT KIND OF GAME YOU WANNA PLAY?</h2>
 				<div class="buttons-container">
-					<button class="button1" onclick="loadContent('onevsone/'); return false;" type="submit">1 VS 1</button>
-					<button class="button2" onclick="loadContent('rumble/'); return false;" type="submit">RUMBLE</button>
+					<button class="button1" onclick="loadContent('pong/onevsone/'); return false;" type="submit">1 VS 1</button>
+					<button class="button2" onclick="loadContent('pong/rumble'); return false;" type="submit">RUMBLE</button>
 				</div>
 			</div>
 		</div>
