+++ transcendence/base/templates/rumble.html
@@ -48,7 +48,7 @@
 					type="text"
 					name="player2"
 					/>
-					<button class="button3" onclick="loadContent('play/'); return false;" type="submit">OK</button>
+					<button class="button3" onclick="loadContent('pong/play/'); return false;" type="submit">OK</button>
 				</div>
 			</div>
 		</div>
@@ -78,7 +78,7 @@
 						type="text"
 						name="player3"
 						/>
-						<button class="button3" onclick="loadContent('play/'); return false;" type="submit">OK</button>
+						<button class="button3" onclick="loadContent('pong/play/'); return false;" type="submit">OK</button>
 					</div>
 				</div>
 				<div class="input-group player2-container">
@@ -106,12 +106,12 @@
 						type="text"
 						name="player4"
 						/>
-						<button class="button3" onclick="loadContent('play/'); return false;" type="submit">OK</button>
+						<button class="button3" onclick="loadContent('pong/play/'); return false;" type="submit">OK</button>
 					</div>
 				</div>
 			</div>
 		<div class="button3-container">
-			<button class="button4" onclick="loadContent('playrumble/'); return false;" type="submit">LET'S FIGHT!!!</button>
+			<button class="button4" onclick="loadContent('pong/playrumble/'); return false;" type="submit">LET'S FIGHT!!!</button>
 		</div>
 	</div>
 </div>
