+++ transcendence/transcendence/urls.py
@@ -25,15 +25,10 @@ urlpatterns = [
     path("otp/", custom_auth_views.otp_view, name="otp"),
     path("verify-otp/", custom_auth_views.verify_otp, name="verify_otp"),
     path("admin/", admin.site.urls, name="admin"),
-    path('user/', include('user.urls')),
+    path("user/", include("user.urls")),
+    path("pong/", include("pong.urls")),
     path('settings/', settingsView, name='settings'),
     path("", base_views.index, name="index"),
-    path("play/", base_views.play, name="play"),
-    path("onevsone/", base_views.onevsone, name="onevsone"),
-    path("rumble/", base_views.rumble, name="rumble"),
-    path("playonevsone/", base_views.playonevsone, name="playonevsone"),
-    path("playrumble/", base_views.playrumble, name="playrumble"),
-    path("tournaments/", base_views.tournaments, name="tournaments"),
     path("play/create/", CreateGame.as_view(), name="game-create"),
     path("play/<int:game_id>/", GameGetOne.as_view(), name="game-detail"),
     path("play/games/", GameGetAll.as_view(), name="games-all"),
